# src/server/openapi/hermes-api.yaml
openapi: 3.1.0
info:
  title: Hermes API
  description: |
    HTTP API für die Hermes LAN-Party-Koordination.

    **Authentifizierung**
    - **Browser:** Session-Cookie `hermes_session` nach Login (wie die Web-UI).
    - **API-Client:** `Authorization: Bearer <token>` mit einem im Profil erzeugten API-Token.

    **CSRF:** Schreibende Anfragen aus dem Browser senden den Header `x-hermes-csrf` (siehe `GET /api/auth/csrf`).
    Bei reiner Bearer-Authentifizierung entfällt die CSRF-Pflicht.

    **API-Token-Scopes**
    - `full`: dieselben Rechte wie der zugehörige Benutzer (Admin/Manager/…); Schreibzugriff wie mit Session.
    - `read_only`: nur `GET`/`HEAD`/`OPTIONS`; schreibende Methoden liefern `403` mit `api_token_nur_lesen`.

    **Hinweis:** Ist der Header `Authorization: Bearer …` gesetzt, wird bei ungültigem Token **kein** Session-Cookie
    als Fallback verwendet (explizit nur Cookie *oder* gültiger Bearer).

    Öffentlich ohne Login: `GET /api/health`, `GET /api/settings/public`, `GET /api/openapi.yaml`, `GET /api/docs`, Kiosk-Pfad (falls aktiviert).
  version: 0.7.0
servers:
  - url: /
    description: Relativ zur Hermes-Instanz
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/events:
    get:
      tags: [events]
      summary: Alle Events (nicht soft-deleted)
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/GameEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [events]
      summary: Event anlegen
      description: Benötigt Manager-, Organisator- oder Admin-Rolle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreate"
      responses:
        "201":
          description: Angelegt
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/events/{id}:
    patch:
      tags: [events]
      summary: Event aktualisieren
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Event abgeschlossen
  /api/events/{id}/participation:
    post:
      tags: [events]
      summary: Teilnahme setzen (dabei / nicht dabei)
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: u.a. Event voll
  /api/events/{id}/cancel:
    post:
      tags: [events]
      summary: Event stornieren
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "204":
          description: OK
  /api/events/{id}/archive:
    post:
      tags: [events]
      summary: Event archivieren
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "204":
          description: OK
...
  /api/admin/events/{id}:
    delete:
      tags: [admin]
      summary: Event soft-deleten (Admin)
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "204":
          description: OK
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: hermes_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API-Token
  parameters:
    EventId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Nicht angemeldet
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: nicht_angemeldet
    Forbidden:
      description: Fehlende Rechte oder read-only-Token auf Schreiboperation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            admin:
              value:
                error: admin_erforderlich
            readOnly:
              value:
                error: api_token_nur_lesen
    BadRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Maschinenlesbarer Code
...
    GameEvent:
      type: object
      properties:
        id:
          type: string
        gameTitle:
          type: string
        startMode:
          type: string
          enum: [now, scheduled]
        startsAt:
          type: string
        minPlayers:
          type: integer
        maxPlayers:
          type: integer
        details:
          type: string
          nullable: true
        status:
          type: string
        joinedCount:
          type: integer
        myParticipation:
          type: string
          nullable: true
          enum: [joined, declined]
    EventCreate:
      type: object
      required: [gameTitle, startMode, minPlayers, maxPlayers]
      properties:
        gameTitle:
          type: string
        startMode:
          type: string
          enum: [now, scheduled]
        startsAt:
          type: string
          description: ISO-8601, Pflicht wenn startMode scheduled
        minPlayers:
          type: integer
        maxPlayers:
          type: integer
        details:
          type: string
          maxLength: 4000